requirements.txt
python-telegram-bot
sentence-transformers
ollama
python-dotenv
numpy
